refactor send command
