+++ packages/expo-cli/src/credentials/utils/git.ts
